+++ src/vrtgen/parser/__init__.py
@@ -22,7 +22,9 @@ import logging
 
 import yaml
 
-from .packet import DataPacketParser, ContextPacketParser, CommandPacketParser
+from vrtgen.model.config import create_packet, PacketType
+
+from .packet import create_parser
 
 __all__ = (
     'parse_packet',
@@ -37,19 +39,21 @@ def parse_packet(name, value):
     if not isinstance(value, dict):
         raise RuntimeError('Invalid definition for packet ' + name)
 
+    log = logging.getLogger(name)
     packet_type = value.pop('type', None)
     if packet_type is None:
         raise RuntimeError('No packet type specified for ' + name)
-    if packet_type == 'data':
-        parser = DataPacketParser(name)
-    elif packet_type == 'context':
-        parser = ContextPacketParser(name)
-    elif packet_type == 'command':
-        parser = CommandPacketParser(name)
-    else:
-        raise RuntimeError("Invalid type '{0}' for packet '{1}'".format(packet_type, name))
+    if packet_type == 'command':
+        log.warning('"command" packet type is deprecated, using "control"')
+        packet_type = 'control'
 
-    return parser.parse(value)
+    packet_type = PacketType(packet_type.casefold())
+    packet = create_packet(packet_type, name)
+    parser = create_parser(packet_type)
+
+    parser(log, packet, value)
+
+    return packet
 
 def parse_file(filename):
     """
@@ -73,5 +77,7 @@ def parse_stream(stream):
                 packet = parse_packet(name, value)
                 packet.validate()
                 yield packet
+            except ValueError as exc:
+                logging.error("%s", str(exc))
             except RuntimeError as exc:
                 logging.exception("%s %s", name, str(exc))